pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repoai"
description = "AI-assisted repository content creation and edition"
authors = [{name = "Carlos Gaete-Morales", email = "cdgaete@gmail.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "gitpython",
    "appdirs",
    "chardet",
    "docker",
    "litellm",
    "rich"
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest",
    "flake8",
    "mypy",
    "black",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["repoai*"]
namespaces = false

[tool.setuptools.dynamic]
version = { attr = "repoai.__version__" }

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.black]
line-length = 150
target-version = ['py39']

[project.scripts]
repoai = "repoai.main:main"
